[feat] admin 테이블 스키마 추가 및 기능 구현
