+++ prisma/schema.prisma
@@ -8,24 +8,50 @@ datasource db {
 }
 
 model User {
-  id           Int      @id @default(autoincrement())
-  email        String   @unique
+  id           Int       @id @default(autoincrement())
+  email        String    @unique
   password     String?  
   nickname     String
-  profileImage Int      @default(0)
-  createdAt    DateTime @default(now())
-  updatedAt    DateTime @updatedAt
+  profileImage Int       @default(0)
+  grade        UserGrade?
+  createdAt    DateTime  @default(now())
+  updatedAt    DateTime  @updatedAt
 
   oauthAccounts OAuthAccount[]
 
   @@map("users")
 }
 
+model Admin {
+  id        Int       @id @default(autoincrement())
+  email     String   @unique
+  password  String?
+  nickname  String
+  role      AdminRole @default(ADMIN)
+  createdAt DateTime  @default(now())
+  updatedAt DateTime  @updatedAt
+
+  @@map("admins")
+}
+
+enum AdminRole {
+  ADMIN
+}
+
+enum UserRole {
+  USER
+}
+
+enum UserGrade {
+  NORMAL
+  EXPERT
+}
+
 model OAuthAccount {
   id           Int      @id @default(autoincrement())
   provider     String   // "google"
   providerId   String   // Google sub
-  email        String?  // 구글 이메일(참고용)
+  email        String?  // 구글 이메일
   userId       Int
   user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
   createdAt    DateTime @default(now())
